backend/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
channels==4.0.0
django-filter==22.1
djangochannelsrestframework==1.1.0
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-yasg==1.21.4
pytest==7.2.0
pytest-django==4.5.2
pytest-asyncio==0.20.3
channels_redis==4.0.0
Pillow==9.3.0
django-cors-headers==3.13.0
django-webpack-loader==1.8.0
psycopg2-binary==2.9.5
django-environ==0.9.0
